Try to get adversarial unit test to work
I tried vectorizing 8 environments to increase the amount of data
available to the generator. Interestingly, this resulted in no
change in run time! (Iterations per second remained the same at
about ~1iter/sec.)

These time results are evidence suggesting that the bottleneck is
the _sess.run call in reward evaluation.