test_airl: Don't cache rollouts during GAIL tests
When checking that training the generator increases the discriminator
loss, caching rollouts was okay in the AIRL case because the
loss is a function of the generator action probabilities.

However, in the GAIL case, the discriminator loss is only a function of
the generator through state-actions chosen by the generator rollout.
Consequently, the discriminator loss stays exactly the same when we
trained the generator but cached the previous rollout. (Leading to test
fail.)