BC: Add n_batches end condition, modify callback
Adds `n_batches` parameter to BC.train(), as requested by @qxcv,
while retaining the `tqdm` progress bar.

Refactoring changed the location in the code where an epoch end
can be straightfowardly detected, so the callback was moved inside
a different function. Accordingly, I removed the `locals` argument
to the callback, which probably wasn't great form in the first place.

We could consider using SB3-style callbacks in the future.