make_vec_env: Make RolloutInfoWrapper optional
This wrapper increases the size of the info dict substantially
because it appends all episode observatations to the into dict with
done=True. It is only used by the expert_demos script to save
original (unnormalized) observations and rewards, so this commit
modifies `make_vec_env` to only optionally wrap each Env with
`RolloutInfoWrapper`.