Don't set policy's env in generate_trajectories (#299)
Because generate_trajectories only uses policy.predict,
the environment doesn't need to be set. Doing so anyway
was an unnecessary side effect that might be undesirable
in some circumstances. The only relevant effect of
policy.set_env was checking whether the action and observation
spaces match, this is now done directly.