Replace hasattr checks
This was very brittle if someone calls the logger e.g. self._logger
in other implementations. Now the logger is set in the base class.
This necessitated introducing ABCs for Fragmenter etc., which I
originally wanted to avoid, but I think it starts to make sense.
I also did this for PreferenceGatherer though it currently doesn't
use the logger because I strongly assume that it or at least
other implementations will at some point.