Implement preference comparisons (#320)
* Generalize RewardNet base class

The previous RewardNet contained some things that are specific to
AIRL or at least not necessary for some other applications
(in particular the train/test split).

* Fix typo in docstring

Co-authored-by: Steven H. Wang <wang.steven.h@gmail.com>

* Remove pass statement in abstractmethod

This is consistent with the rest of the codebase and avoids
codecov complaints.

* Reduce AIRL-specific reward net code

Almost all of the code in reward_nets is now completely general
and can be used by other algorithms in the future.
Only AIRLRewardNet is specific to AIRL (and it's simply a thin wrapper).

* Add test for RewardNet.device without parameters

* Remove pass statement to avoid codecov issues

* Add first preference comparisons prototype

* Fix line lengths

* Add test and fix lots of bugs

* Fix logger error in test

* Fix docstring formatting issue

* Improve doc string

Co-authored-by: Adam Gleave <adam@gleave.me>

* Improve type signatures

Co-authored-by: Adam Gleave <adam@gleave.me>

* Fix missing imports and create type aliases

* Fix imports to adhere to Google style

Not for imitation.data.types for now, will decide on that later.

* Fix another import

* Sample inside SyntheticGatherer

Instead of returning analytic probabilities, we now sample
discrete decisions (0 or 1).

* Add discount factor to SyntheticGatherer

* Improve doc string

Co-authored-by: Adam Gleave <adam@gleave.me>

* Simplify code slightly

* Slight refactoring

* Fix docstrings

* Add assert statements

* Add discounting in reward model training

I think this makes sense given that we now support
discounting for the synthetic gatherer.

* Fix dtype

* Add another assert

* Split up train() and sample()

* Add TODO comment

* Decrease warning threshold and improve docstring

* Add logging and training script (still WIP)

* Apply suggestions from code review

Co-authored-by: Adam Gleave <adam@gleave.me>

* Fix many small issues from code review

* Save final models

* Expose more arguments via CLI

* Evaluate trained agent at the end

* Rename "steps" to "iterations"

I think this makes it clearer that his is "one level higher"
than the agent training steps.

* Speed up smoke test

* Move Fragmenter into preference_comparisons

Seems unlikely to be used elsewhere and doesn't
make much sense to have inside imitation.data.

* Change some default hparams

These aren't tuned but at least they quickly solve Cartpole.

* Train reward model before agent

Training the agent on the initially random reward model
in the first iterations doesn't make sense.

This also required sampling trajectories if none have been
gathered yet.

* Increase test coverage

* Upgrade SB3 and make logger.py non-global; much breakage elsewhere

* Make BC use local logger; if folder None default to temporary directory

* Lint

* Make adversarial trainer and DAgger local logger

* Make DAgger and BC pickleable again

* Remove old workarounds for SB3 issue #109

* 4 TB directories expected (we reduced to 2 only because SB TB disabled previously)

* Boost test coverage

* Exempt coverage of intentionally dead test code

* Rename _HierarchicalLogger in docstring

* Switch to temperature param and improve coverage

* Add more tests

* Switch to local logger

* Configure logger correctly in script

* Remove duplicate log message

* Replace hasattr checks

This was very brittle if someone calls the logger e.g. self._logger
in other implementations. Now the logger is set in the base class.
This necessitated introducing ABCs for Fragmenter etc., which I
originally wanted to avoid, but I think it starts to make sense.
I also did this for PreferenceGatherer though it currently doesn't
use the logger because I strongly assume that it or at least
other implementations will at some point.

* Remove unnecessary warn method

* Delete unused import

* Simplify tests

* Ignore unused dummy function for coverage

Co-authored-by: Steven H. Wang <wang.steven.h@gmail.com>
Co-authored-by: Adam Gleave <adam@gleave.me>