Add a program to generate commands to run training scripts with different configs.
See the header file of experiments/commands.py for details. The primary motivation is to make it easy to run the training scripts using the config files in benchmarking/. It attempts to simplify (1) running local commands in parallel compared to the bash scripts in experiments/ and (2) running commands in separate containers on the Hofvarpnir cluster compared to parallel.py, which relies on Ray. AdamGleave@'s thoughts on using Ray on Hofvarpnir: "We could make Ray work I think, but it'd be janky. It'd be fairly easy to e.g. have a Docker container that just starts a Ray server, then the command runs on that Ray server, and the whole container ends when the job is complete. But then you only get parallelism up to the max machine size, no distributed training. The other option is to just launch a bunch of Ray server jobs up to the expected size, then run other jobs that connect to those Ray servers. But then you'd have to teardown the Ray servers manually to avoid bleeding resources. Overall I think it probably does make sense to find a non-Ray solution for your part of the project, though I still feel moderately excited about getting a Ray solution working longer-term. It's probably not that hard to make the Ray autoscaler handle this."