Enclose the training command within the cluster command in double quotes and escape double quotes in the training command.
Previously, we did not enclose the training command in quotes and instead just escaped spaces in the training command. This makes the training command less readable.

This commit also add some tests. In particular, it modifies the test_commands_local_config_runs to check that the expected output directories exist. And it adds 2 tests related to special characters like a space appearing in the flag values passed to commands.py.

If we want to use a space in the output_dir, for example, we can enclose the directory in escaped double quotes and enclose all that in double quotes as follows:

```bash
python commands.py \
	--name=run0 \
	--cfg_pattern=../benchmarking/example_bc_seals_half_cheetah_best_hp_eval.json \
	--output_dir="\"foo bar\"" | bash
```

And similarly when running with the --remote flag:

```bash
python commands.py \
	--name=run0 \
	--cfg_pattern=../benchmarking/example_bc_seals_half_cheetah_best_hp_eval.json \
	--output_dir="\"/data/foo bar\"" \
	--remote
```

I tested the commands above run and write out the expected directories.