Fix benchmarking configs + Add a minimal test suite for those configs. (#653)
* Fix benchmarking configs + Add a minimal test suite for those configs.

This commit patches in the changes from PR #650. It also fixes an error in the dagger configs by replacing "py/object" with "py/type". "py/object" failed the assert in sacred.config.utils.assert_is_valid_key. Finally, it adds a minimal test suite for the benchmarking configs. This commit makes progress on testing, but still does not test that all the benchmarking configs work. Running all the configs with the fast configs applied is too slow. Calling sacred's print_config command for all the configs is also too slow.

* Update the tests for benchmarking configs so that they do not require MuJoCo.

* Remove unused global in test_benchmarking.py

* Test all the benchmarking configs using print_config.

Avoid the overhead of subprocess by running the print_config using the experiment's run method. Also, use a path relative to the script for BENCHMARKING_DIR instead of creating a path and then checking if it exists.

* Minor fixes

Co-authored-by: Adam Gleave <adam@gleave.me>