Add a program to generate commands to run training scripts (#654)
* Add a program to generate commands to run training scripts with different configs.

See the header file of experiments/commands.py for details. The primary motivation is to make it easy to run the training scripts using the config files in benchmarking/. It attempts to simplify (1) running local commands in parallel compared to the bash scripts in experiments/ and (2) running commands in separate containers on the Hofvarpnir cluster compared to parallel.py, which relies on Ray. AdamGleave@'s thoughts on using Ray on Hofvarpnir: "We could make Ray work I think, but it'd be janky. It'd be fairly easy to e.g. have a Docker container that just starts a Ray server, then the command runs on that Ray server, and the whole container ends when the job is complete. But then you only get parallelism up to the max machine size, no distributed training. The other option is to just launch a bunch of Ray server jobs up to the expected size, then run other jobs that connect to those Ray servers. But then you'd have to teardown the Ray servers manually to avoid bleeding resources. Overall I think it probably does make sense to find a non-Ray solution for your part of the project, though I still feel moderately excited about getting a Ray solution working longer-term. It's probably not that hard to make the Ray autoscaler handle this."

* Clean up commands.py and associated tests.

In particular:
* Use a template for the expected commands for the 2 local config tests
* Use triple quotes for long strings
* Use a path relative to the script instead of creating a path and testing if it exists
* Use nargs=+ for --seeds flag
* Put main code into a main and parse function
* Add additional docstrings
* Don't use a GPU for the cluster commands
* Remove --high-priority for cluster commands
* Add --shared-host-dir-mount flag

* Fix example commands in the header.

* Add a docstring to the parse function.

* Skip commands.py tests on Windows.

See test failures here:
https://app.circleci.com/pipelines/github/HumanCompatibleAI/imitation/3280/workflows/4708172e-7633-475d-bc73-aabbce65c90e/jobs/11919

* Enclose the training command within the cluster command in double quotes and escape double quotes in the training command.

Previously, we did not enclose the training command in quotes and instead just escaped spaces in the training command. This makes the training command less readable.

This commit also add some tests. In particular, it modifies the test_commands_local_config_runs to check that the expected output directories exist. And it adds 2 tests related to special characters like a space appearing in the flag values passed to commands.py.

If we want to use a space in the output_dir, for example, we can enclose the directory in escaped double quotes and enclose all that in double quotes as follows:

```bash
python commands.py \
	--name=run0 \
	--cfg_pattern=../benchmarking/example_bc_seals_half_cheetah_best_hp_eval.json \
	--output_dir="\"foo bar\"" | bash
```

And similarly when running with the --remote flag:

```bash
python commands.py \
	--name=run0 \
	--cfg_pattern=../benchmarking/example_bc_seals_half_cheetah_best_hp_eval.json \
	--output_dir="\"/data/foo bar\"" \
	--remote
```

I tested the commands above run and write out the expected directories.

* Remove look up of USER environment variable in test.

* Update experiments/commands.py

Co-authored-by: Adam Gleave <adam@gleave.me>

* Update experiments/commands.py

Co-authored-by: Adam Gleave <adam@gleave.me>

Co-authored-by: Adam Gleave <adam@gleave.me>